integrated sessions
Sessions are now mainly integrated - I will come back and review later if there are any bugs or security issues with the code.

The Sync() function does not actually make the entire function sycronous - it actually optimizes the thread. Only the .sync() functions are sycronous.

I wanted to really avoid any sync functions, but the username is so essential to the entire data processing that it had to me for a lot of functions.